riverwood-main/riverwood-voice-agent/requirements.txt
flask==3.0.0
openai==1.54.0
elevenlabs==1.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
